// ClientApp/src/Types/ICompanysBooking.ts
export default interface ICompanysBooking {
  clients_clientID: string;
  clients_name: string;
  clients_surname: string;
  clients_idNumber: string;
  clients_phonenumber: string;
  clients_email: string;
  clients_genderID: number;
  clients_dateOfBirth: string;
  clients_titleID: number;
  clients_createdDate: string;
  bookings_bookingID: string;
  bookings_clientID: string;
  bookings_eventClassID: string;
  bookings_paymentMethodID: number;
  bookings_voucher: string;
  bookings_reference: string;
  bookings_amountPaid: number;
  bookings_proofOfPaymentLink: string;
  bookings_ticketCode: number;
  bookings_ticketDescription: string;
  bookings_isReserved: boolean;
  bookings_isBooked: boolean;
  bookings_isCheckedIn: boolean;
  bookings_reservedDate: string;
  bookings_unreservedDate: string;
  bookings_bookedBy: string;
  bookings_bookedDate: string;
  bookings_unbookedBy: string;
  bookings_unbookedDate: string;
  bookings_checkedInBy: string;
  bookings_checkedInDate: string;
  bookings_uncheckedInBy: string;
  bookings_uncheckedInDate: string;
  bookings_editedBy: string;
  bookings_editedDate: string;
  eventsClasses_eventClassID: string;
  eventsClasses_eventID: string;
  eventsClasses_classID: string;
  eventsClasses_name: string;
  eventsClasses_code: string;
  eventsClasses_price: number;
  eventsClasses_fromPlace: string;
  eventsClasses_toPlace: string;
  eventsClasses_ticketsRangedFrom: number;
  eventsClasses_ticketsRangedTo: number;
  eventsClasses_totalNumberOfTickets: number;
  eventsClasses_availableNumberOfTickets: number;
  eventsClasses_bookedNumberOfTickets: number;
  eventsClasses_isActive: boolean;
  eventsClasses_activatedBy: string;
  eventsClasses_activatedDate: string;
  eventsClasses_deactivatedBy: string;
  eventsClasses_deactivatedDate: string;
  eventsClasses_descriptions: string;
  eventsClasses_createdBy: string;
  eventsClasses_createdDate: string;
  eventsClasses_editedBy: string;
  eventsClasses_editedDate: string;
}